C++ Project - 1 - Battleship
Learning Objectives:
●	Write C++ programs in a modern software development using the CLion IDE and version control with Github repositories.
●	Perform basic computational operations such as i/o, 
●	Apply control structures such as if statements, loops, and error messages.
●	Implement and manipulate the contents of a 2D vector.

For Project 1, you will complete a C++ program in CLion and use a Github repository where you will “push” your C++ program (see Appendix).


Battleship

Write a program named “Battleship.cpp” that will print the matrix 3 x 3 from the user and repeatedly ask the user to guess a location on the matrix. Using a random number generator, you will choose two numbers between 1 and 3, inclusive, and mark that location on your 3 x 3 matrix but not show this to the user. Each time the user guesses a location, you will update the location on your matrix, print the matrix, and keep track of how many guesses the user has made. If the user correctly guesses the location, the program will print the updated matrix, output the total number of guesses, and end the program.